Show failed pull operations in red in the UI.
