+++ frontend/src/routes/channels/id/BatchPullCommentsSidebar.tsx
@@ -3,13 +3,12 @@ import {
     Button,
     Col,
     Container,
-    ListGroup,
-    ListGroupItem,
     Offcanvas,
     ProgressBar,
     Row,
 } from "react-bootstrap";
 import { AvailableQuota } from "#/components/limits/AvailableQuota";
+import { PullCommentsList } from "#/components/pull/PullCommentsList";
 import { BatchPullCommentsAction } from "#/routes/channels/id/BatchPullCommentsAction";
 import { useAppDispatch, useAppSelector } from "#/store/hooks";
 import { invalidateVideos } from "#/store/slices/backend";
@@ -18,7 +17,6 @@ import {
     selectAllPullComments,
 } from "#/store/slices/pulls";
 import { pluralize } from "#/utils";
-import type { PullCommentsState } from "#/store/slices/pulls";
 import type { ChannelProps } from "#/types";
 
 const BatchPullCommentsSidebar = ({ channel }: ChannelProps) => {
@@ -81,64 +79,6 @@ const BatchPullCommentsSidebar = ({ channel }: ChannelProps) => {
                 </Row>
             </Offcanvas.Body>
         </Offcanvas>
-    );
-};
-
-type PullCommentsListProps = {
-    pullComments: PullCommentsState[];
-};
-
-const PullCommentsList = ({ pullComments }: PullCommentsListProps) => (
-    <ListGroup variant={"flush"}>
-        {pullComments.map((pullComment, index) => (
-            <PullCommentsItem
-                key={pullComment.video.id}
-                pullComment={pullComment}
-                index={index}
-            />
-        ))}
-    </ListGroup>
-);
-
-type VideoItemProps = {
-    pullComment: PullCommentsState;
-    index: number;
-};
-
-const PullCommentsItem = ({ pullComment, index }: VideoItemProps) => {
-    const started = pullComment.status !== "READY";
-    const active = pullComment.status === "FETCHING";
-
-    const commentsLabel = pluralize(pullComment.commentIds.length, "comment");
-    const repliesLabel = pluralize(
-        pullComment.replyIds.length,
-        "reply",
-        "replies",
-    );
-
-    return (
-        <ListGroupItem variant={active ? "info" : ""}>
-            <Row>
-                <Col className={"line-clamp-1"}>
-                    <strong>{index + 1}.</strong> {pullComment.video.title}
-                </Col>
-            </Row>
-            <Row>
-                <Col>
-                    <ProgressBar
-                        animated={active}
-                        label={`${commentsLabel} and ${repliesLabel}`}
-                        now={started ? 100 : 0}
-                        variant={active ? "info" : started ? "success" : ""}
-                    />
-                </Col>
-            </Row>
-            {/*<Row>*/}
-            {/*    <Col xs={"auto"} className={"small ms-auto"}>*/}
-            {/*        {pluralize(pullComment.commentIds.length, "comment")}*/}
-            {/*    </Col>*/}
-            {/*</Row>*/}
-        </ListGroupItem>
     );
 };
